+++ game/Component/Spinner.hpp
@@ -15,8 +15,8 @@ public:
 		_transform = getComponent<Transform>();
 	}
 
-	void update(float dt){
-		_transform->rotate(Vector3f(0.f, _speed * dt, 0.f));
+	void update(long double dt){
+		_transform->rotate(Vector3f(0.f, (float)(_speed * dt), 0.f));
 
 		const Uint8* keyDown = SDL_GetKeyboardState(0);
 	}
